Refactor service commands a bit
